[DataGrid] Make rowCount part of the state
