+++ docs/pages/x/api/data-grid/data-grid.json
@@ -313,6 +313,10 @@
         "describedArgs": ["params", "event", "details"]
       }
     },
+    "onRowCountChange": {
+      "type": { "name": "func" },
+      "signature": { "type": "function(count: number) => void", "describedArgs": ["count"] }
+    },
     "onRowDoubleClick": {
       "type": { "name": "func" },
       "signature": {
