+++ docs/pages/x/api/data-grid/grid-pagination-api.json
@@ -16,6 +16,11 @@
       "name": "setPaginationModel",
       "description": "Sets the <code>paginationModel</code> to a new value.",
       "type": "(model: GridPaginationModel) => void"
+    },
+    {
+      "name": "setRowCount",
+      "description": "Sets the <code>rowCount</code> to a new value.",
+      "type": "(rowCount: number) => void"
     }
   ]
 }
