+++ packages/grid/x-data-grid/src/components/GridPagination.tsx
@@ -7,9 +7,10 @@ import { styled } from '@mui/material/styles';
 import { useGridSelector } from '../hooks/utils/useGridSelector';
 import { useGridApiContext } from '../hooks/utils/useGridApiContext';
 import { useGridRootProps } from '../hooks/utils/useGridRootProps';
-import { gridFilteredTopLevelRowCountSelector } from '../hooks/features/filter';
-
-import { gridPaginationModelSelector } from '../hooks/features/pagination/gridPaginationSelector';
+import {
+  gridPaginationModelSelector,
+  gridPaginationRowCountSelector,
+} from '../hooks/features/pagination/gridPaginationSelector';
 
 const GridPaginationRoot = styled(TablePagination)(({ theme }) => ({
   [`& .${tablePaginationClasses.selectLabel}`]: {
@@ -35,12 +36,7 @@ export const GridPagination = React.forwardRef<unknown, Partial<TablePaginationP
     const apiRef = useGridApiContext();
     const rootProps = useGridRootProps();
     const paginationModel = useGridSelector(apiRef, gridPaginationModelSelector);
-    const visibleTopLevelRowCount = useGridSelector(apiRef, gridFilteredTopLevelRowCountSelector);
-
-    const rowCount = React.useMemo(
-      () => rootProps.rowCount ?? visibleTopLevelRowCount ?? 0,
-      [rootProps.rowCount, visibleTopLevelRowCount],
-    );
+    const rowCount = useGridSelector(apiRef, gridPaginationRowCountSelector);
 
     const lastPage = React.useMemo(
       () => Math.floor(rowCount / (paginationModel.pageSize || 1)),
