+++ scripts/x-data-grid-pro.exports.json
@@ -363,6 +363,7 @@
   { "name": "GridPaginationInitialState", "kind": "Interface" },
   { "name": "GridPaginationModel", "kind": "Interface" },
   { "name": "gridPaginationModelSelector", "kind": "Variable" },
+  { "name": "gridPaginationRowCountSelector", "kind": "Variable" },
   { "name": "gridPaginationRowRangeSelector", "kind": "Variable" },
   { "name": "gridPaginationSelector", "kind": "Variable" },
   { "name": "GridPaginationState", "kind": "Interface" },
